+++ src/tb_watcher/pages.py
@@ -112,8 +112,9 @@ class TwitterThread(TwitterPage):
     Also used to resolve Tweet ID as a thread will contain
     the Tweet's ID.
     """
-    def __init__(self, main_tweet_data: Tweet, *args, **kwargs):
+    def __init__(self, main_tweet_data: Tweet, prior_tweet_data: Tweet, *args, **kwargs):
         self.main_tweet_data = main_tweet_data
+        self.prior_tweet_data = prior_tweet_data
         # Force auto fetching threads to false to prevent recursion.
         super().__init__(*args, **kwargs)
 
@@ -134,9 +135,18 @@ class TwitterThread(TwitterPage):
         tweets = self.driver.find_elements(By.CSS_SELECTOR, '[data-testid="tweet"]')
 
         main_tweet = None
+        parent_id = None
         # Not every tweet on the metadata we want. Keep iterating until we find it.
         for t in tweets:
             dtm = tweet_dom_get_basic_metadata(t)
+            # Check if this tweet is a parent, if so, we can mark it as a parent tweet.
+            if self.prior_tweet_data and \
+               dtm.name == self.prior_tweet_data.name and \
+               dtm.tweet_text == self.prior_tweet_data.tweet_text:
+               print("MATCH")
+               parent_id = self.prior_tweet_data.id
+               print(parent_id)
+
             # Some forms don't exist in thread.
             if dtm.name == self.main_tweet_data.name and \
                dtm.tweet_text == self.main_tweet_data.tweet_text:
@@ -154,6 +164,8 @@ class TwitterThread(TwitterPage):
         post_url =path.split("status/")[-1]
         dtm.id = post_url.split("/")[0]
 
+        # Set parent_id if we found it.
+        dtm.parent_id = parent_id
         self.metadata = dtm
 
         # Create a folder to house pictures, etc.