+++ CHANGELOG.md
@@ -21,10 +21,12 @@
 ## 0.6.0: Twitter Thread Support
 
 * Added `--depth` and `-d` for archiving threads.
-* Added `TwitterPageInterface` used for representing a page in Twitter.
+* Added `TwitterPage` used for representing a page in Twitter.
 * Added `TwitterBio` which is used for fetching and obtaining bio pages.
-* Added `TweetFetcher` which attempts to fetch all available Tweets on a page.
+* Added `TwitterThread` which represents a tweet thread page.
+* Added `TweetExtractor` which attempts to fetch all available Tweets on a page.
 * Added `Scroller` which abstracts away scrolling metrics.
+* Added thread expansion functions: `hit_more_replies()` and `get_recommend_tweets_height()`.
 * Changed random messages from ChromeDriver, makes logs cleaner.
 * Fixed potential for some posts to be skipped on ad removal.
 * Fixed logger not outputting any info.